Fake parton momentum in xml
Added an entry in xml for hadronization
Added functionality in HadronizatoinModuleTest to read this xml file in
::init()

Added the variable p_fake
this is set by the value in the xml file and used to set the momentum
of the fake parton.

Please dont set the momentum of the fake parton in the hadronization
module anymore. Set it in the xml file.