delete termination command when prob > 1
The typical value of deltaT used in martini is around 0.01 fm.
The total probability of having all processes = sum of all rates times deltaT
Thus if you are getting the probability > 1, you need to decrease deltaT.

I commented out the lines for terminating the code when the probabilities for quarks and gluons exceed 1.
I think that a couple of times of getting Prob > 1 doesn't affect overall physics. However if you see a bunch of this error, you should decrease deltaT in the xml file.