Fix PYTHIA_XML_DIR in SMASH
That is likely a SMASH bug, but it did not show up anywhere
and was not changing any results. I am creating a SMASH
issue about it, but here in JetScape, I prefer to fix it
immediately. If I did something wrong, this change will be
overwritten by updated SMASH version.

Signed-off-by: Dmytro Oliinychenko <doliinychenko@lbl.gov>