Energy non-conservation issues fixed.
Defined t = E^2 - px^2 -py^2 -pz^2 - m^2 in JetScapeParticles.cc
Typos in MATTER.cc formulas fixed.
Removed 2 attachments of TRENTO in bricktest.cc, was leading to memory explosion.
Added a pythia helper in PGun to get the mass of the particle and be consistent with JetScapeParticleBase
Initialized tranport coefficients in MATTER