Bugfix: reset SMASH event index properly
For each hydro event there can be N sampler events. Particles from all N
sampler events are saved in a structure, which SMASH wrapper passes to SMASH.
Next hydro event executes new N sampler events and resets this structure, so
particles from previous hydro event are forgotten. SMASH wrapper didn't know
that and incremented the event number further instead of resetting it to 0 for
each new hydro event. That number was used as an array index, so for the second
hydro event it segfaulted.

This refs #55

Signed-off-by: Dmytro Oliinychenko <doliinychenko@lbl.gov>