Proof of concept of SMASH afterburner module running with timesteps
Instead of the whole event running, SMASH only runs its evolution synced
with the main clock. Some small modfications to the SMASH source code
are necessary for this to work. Also, choose smaller timestep in SMASH
than for Jetscape clock in order for it to work propely at the moment.

This version runs with a fixed test hadron list.