Update SMASH Wrapper for changed SMASH functions
In particular, make the end time a member of the wrapper class as it 
needs to be passed explicitly to the run time evolution now.

SMASH also gained the feature to use parallel ensembles. This is not 
forseen to be emplyoed in JETSCAPE, so we only use the first ensemble.