Update to adjust internally how the binning is processes.
From the view of the x-axis: all binning starts at -grid_max_x, and then increases as
nx*step_size_x. To get a symmetrix axis, updated to add one extra bin.

This is also true in MUSIC, that reads the grid size to figure out nx.

Updated to also allow y to update; in some cases just assumed to be symmetrix with x.